[AutofillDataModel] Set up i18n API baseline for address models.
The CL introduces the first set of per country rules for dynamic
hierarchies. It also creates the initial skeleton for the API that will
provide clients with access to international address models.

In the initial setup, the API supports fetching a per-country address
model that is built at runtime from the set of rules. Only a small list
of migrated countries is currently supported.

Bug: 1464568
Change-Id: I656e7cf53eb7c59758e49cd236e5ca560166141a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4705551
Reviewed-by: Dominic Battre <battre@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Norge Vizcay <vizcay@google.com>
Cr-Commit-Position: refs/heads/main@{#1176242}