Hide media metadata from media controllers if required
This CL implements the ability for embedders to obscure the media
metadata from the OS's media controllers. This media metadata will
simply be replaced by some placeholder metadata.

This is done by making use of the embedder-specific MediaSessionClient
API to both check if we need to obscure the media metadata and also get
the placeholders we should replace the media metadata by.

As in Chromium's case the implementation of the MediaSessionClient API
(ChromeMediaSessionClient) specifies that the media metadata should be
hidden if its playing in an Incognito session, that will be the new
behavior introduced by this CL.

The actual obscuring of the metadata is just implemented by
intercepting calls to the SystemMediaControls API and replacing the
metadata with placeholder metadata when needed.

Bug: 1447545
Change-Id: If8fdf1703cb4eda4652f17d1fdd515567a5dcfde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4651207
Reviewed-by: Chris Bookholt <bookholt@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Filipa Senra <fsenra@google.com>
Commit-Queue: Manuel Palenzuela Merino <manuelpalenzuelamerino@gmail.com>
Reviewed-by: Zaina Al-Mashni <zalmashni@google.com>
Cr-Commit-Position: refs/heads/main@{#1176245}