[Static Layout] Fix doneHiding()
In StaticLayout call super.doneHiding() at the end of the override
rather than the beginning. In some cases doneHiding() may re-show the
layout causing mIsActive and the TAB_ID to be in an incorrect state.

A regression due to resetting the TAB_ID was introduced in
https://chromium-review.googlesource.com/c/chromium/src/+/4663171

It is surprising that the issue with mIsActive was not a source of
problems in the past. This might explain some flickering behavior
seen previously.

Demo of fix (compare to video for first bug):
https://drive.google.com/file/d/1lhbMybO7-9UuxZz5aSlQ8Zz8Kp03em_u/view?usp=drive_link

See first bug for a reliable repro. Second bug is a possible repro
that is very occasional, but has similar characteristics.

As a follow up I'll look to add a test for this case. Unfortunately,
the reliable repro relies on self-sharing through the Android
share sheet which isn't easy to replicate in a test environment.

Bug: 1468214, 1467884
Change-Id: I28a6dcf38b78ef9fa69ea2ffe08d7e20df3f8ed5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4725612
Commit-Queue: David Trainor <dtrainor@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176261}