[cookie controls] Record a setting when the entry point was animated
Introduce cookie controls metadata website setting. It contains metadata
related to creating exceptions through cookie controls UI, like: was
the entry point already animated for the site, last expiration data of
the exception (not in this CL), repeated activations count (not in
this CL).

When a high confidence signal is reported, record that in the setting.
If the setting is set, the high confidence signal won't be reported for
that site anymore.

Bug: 1446230, b:280782678
Change-Id: Idbbc18fe14f05cfb75d9609fe611df8c03d3bbde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4701424
Owners-Override: Krishna Govind <govind@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Krishna Govind <govind@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/main@{#1176277}