[Auto-disable Accessibility] Add UMA histograms for web contents accessibility
This CL continues the Auto-disable Accessibility feature work.

With this CL we add three new histograms that will track the usage of
the accessibility engine as it compares to the general time usage of
a given Tab/WebContents instance. We track the time that a WebContents
object is shown and hidden, and record the time between the two (this
can happen more than once per Tab). We also track when the native
accessibility code is lazily initialized (which may not be the same as
when the WebContents was shown), and record the time since then when a
WebContents is hidden. In the special case where these two times are
nearly the same (sans some noise), we record the time in a separate
histogram. This will allow us to compare the average time that
accessibility services are enabled in general, vs. the amount of time
that accessibility services are enabled for users that enable the
services 100% of the time.


AX-Relnotes: N/A
Bug: 1430202, b/265493191
Change-Id: I673f8b5727fff785ab9f3ba54a36d38c1d52a950
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4706329
Reviewed-by: Amanda Lin Dietz <aldietz@google.com>
Reviewed-by: Evan Liu <evliu@google.com>
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1176289}