[iOS] Handle newly added overflow menu actions
This CL checks if any of the available actions are not currently stored
in either the shown or hidden actions list (i.e. are new) and adds them
to the shown list if so. This also neatly handles the new user case.
New users have nothing stored in shown or hidden, so everything will
be added to shown on the first run.

Bug: 1445203
Change-Id: Ib375116d65bf70cb8f5a3fda8a76983ba6885f33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4726445
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Nicolas MacBeth <nicolasmacbeth@google.com>
Cr-Commit-Position: refs/heads/main@{#1176290}