[ios] Support new "only button" style for sign-in promo card
No behavior change because the support is added but not used yet. Add
a new SigninPromoViewStyle that only displays the primary button with
text "Turn On"; no secondary button, avatar, or label. This is done to
satisfy a weird construct in Recent Tabs (https://crbug.com/1467810).

The new style will only be used when kReplaceSyncPromosWithSignInPromos
is enabled. The current code directly manipulates SigninPromoView and
fixing that is not a goal, we'll just wait until the flag is launched
then removed.
https://source.chromium.org/chromium/chromium/src/+/main:ios/chrome/browser/ui/recent_tabs/recent_tabs_table_view_controller.mm;l=1186-1187;drc=1761a16b35b8afcf6b7b73e80c8f437d80832dac

Unrelated to this, we also clean up one no-op here.
https://source.chromium.org/chromium/chromium/src/+/main:ios/chrome/browser/ui/authentication/cells/signin_promo_view.mm;l=596;drc=0ce365a563546008894a3d9a0a207cd41c820adf

Bug: 1447012
Change-Id: Ifc888ce5e50eb143425a2d39c2092667e2f1060d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4725363
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Auto-Submit: Victor Vianna <victorvianna@google.com>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#1176319}