[iOS] Move unopened_downloads_tracker to a different file
Keeping track of the unopened download is independant of the download manager coordinator. Indeed, Chrome should keep track even after the manager is dismissed.
In order to ensure the download manager coordinator is stopped when its view is dismissed, it means another class should own the unopened download tracker. Hence this class should be able to access it. Hence we need to declare it in the header. Let’s move it to a different file at the same time.

Bug: 1467048
Change-Id: Ie009efd5fcfc690dd9134e7d6549aea129b66fc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4708869
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176321}