[ios/uno] Integrate Recent Tabs with sign-in sheet
Behavior changes are behind flag. Have the promo card in recent tabs
show the sign-in sheet instead of the sync consent screen.

For next CLs:
- Show the history opt-in screen after the sign-in sheet. And sign the
user out if the reject that second screen.
- Show the history opt-in screen if the user interacts with the promo
when they are signed-in but not opted in to history sync. For now the
behavior is the old one, i.e. open the UI with the data type toggles.

Bug: 1447012, 1447014, 1466884
Change-Id: I6e5fb2c08d5e52354e06b2a148cef031dce7eca3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4634139
Auto-Submit: Victor Vianna <victorvianna@google.com>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176322}