Fix Race with XrSessionCoordinator destruction
With both Cardboard and ArCore using the XrSessionCoordinator, it is
possible for there to be a race between one device tearing down a
session and the other being destroyed, resulting in one device trying
to destroy their instance of the XrSessionCoordinator while the other
device has an ongoing session. In this case sActiveSessionInstance would
not be null as we previously asserted. This change fixes the assertion
to more accurately reflect what we want which is that we are not trying
to destroy an XrSessionCoordinator that is responsible for an active
session.

Change-Id: I5e9951801b2835dc67f53a95c8b0dadb2da77cdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727082
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Piotr Bialecki <bialpio@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176323}