Polish how glanceables bubble container resizes
Track when the prefererd size of individual glanceable bubble views
changes, new glanceables view get added, or when glanceables view
visibility change and update the glanceable tray bubble bounds in each
of those cases. This ensures that the glanceable tray bubble bounds grow
as new glanceable views get added to the scrollable bubble container
(which happens asynchronously after the bubble is first shown).

Additionally, when the glanceable views container height changes (e.g.
because a new glanceable view gets added, or a glanceable view's size
changes upon selecting a different list), update the glanceables tray
bubble view scroll view offset so it appears that the glanceables UI is
growing upwards (i.e. try to keep the bottom of the scroll view
constants). As special case, ensure that the currently focused view is
scrolled to be visible, to reduce jankiness when the user selects a list
that causes the glanceable to grow and move the dropdown control out of
the screen bounds.

BUG=b:293391626

Change-Id: I90153108b71509991922fde8425ea6f4e43ff9b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4723265
Reviewed-by: Matthew Mourgos <mmourgos@chromium.org>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176335}