Ignore launching missing pWA
When launching a floating workspace from another device onto a
device that do not have PWAs in the template, it would launch them
in a browser window, but does not actually fully load and the
shelf icons are missing. When attempting to save the desk as a
template, it would crash. We should ignore the missing PWA.

BUG=b:293477863
TEST=manually,DeskClients browser test

Change-Id: I62a783842cae297f193eafe4baf7b8c364db0bea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4725245
Commit-Queue: Matthew Zhu <zhumatthew@google.com>
Reviewed-by: Daniel Andersson <dandersson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176339}