Node connected check should also check owner shadow hosts
We should not create AXObjects for, or process changes for nodes that are not "connected".

A node can return true for node->isConnected(), but be inside of a
shadow root, where the shadow host is not connected. This CL replaces the a11y code's isConnected() checks with a helper that determines whether a node and all of its shadow hosts are connected.

Bug: none
Change-Id: I147a4594e0edee69c3bfe83e0237d4d1e20e1755
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4711192
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176347}