[Extensions] Remove NOTREACHED in the Bookmarks API.
BookmarkEventRouter implements BookmarkModelObserver::
BookmarkAllUserNodesRemoved with a comment that this is only called
on Android, not on desktop. However, this is not true, and the
NOTREACHED is being triggered. This CL removes the NOTREACHED and
adds a TODO referencing a new tracking to decide whether we need to
respond to this event in the Bookmarks API.

Bug: 1423472, 1468324
Change-Id: I5f0f27d210f1727d0c0852e2c6b9405a2e0b1461
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4728181
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Solomon Kinard <solomonkinard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176350}