QsRevamp: Bubble should not be closed when a popup close
Fix a bug when a popup is closed, the window activation switches to an
existing window and results in the bubble close. We should lock the
behavior of close bubble on window activated when closing popup, similar
to what we did with the productivity launcher.

Fixed: b:291988617
Change-Id: I3a92a1dacd2d2550810df52da701c92a4661a416
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727586
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Andre Le <leandre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176351}