video-conferece:Update to autohide shelf behavior
Previously autohidden shelfs were shown if a VideoConferenceMediaState
update was sent, but this is not sent when apps begin capturing unless
the media state also happened to update.

Now that HandleClientUpdate was landed, we get poked when a new app
begins capturing. So respect that signal instead.

Change-Id: I31e04c29a4e0c6b7353047d9fe30c7f20faf7d39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4718109
Reviewed-by: Andre Le <leandre@chromium.org>
Commit-Queue: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176361}