Set selection bounds if needed before showing the touch selection menu.
On ChromeOS, tapping the caret should show an insertion handle and
toggle showing the touch selection menu. The bounds of the handle and
menu are set based on the selection bounds of the
TouchSelectionController, however, these selection bounds can be reset
between the time the caret is placed and the time the caret is tapped
again (e.g. if the mouse was moved after placing the caret). In such
cases, tapping the caret makes the menu appear at the wrong spot (top
left of the web content) and the handles don't appear.

To fix this, check if the TouchSelectionController was reset and
re-notify it of the most recently used selection bounds if needed before
showing the menu. This should ensure that the TouchSelectionController
has valid selection bounds when we try to show the handle and menu.

Bug: b:286476233
Change-Id: I1a4ac64f0be8a4f6d096743edc9acc2134cc7903
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4719710
Commit-Queue: Michelle Chen <michellegc@google.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176380}