Reland "[VC] ukm logging for arc++ apps."
This is a reland of commit a3ab73187fdf1bb0fa65df9cfa28d56259f6dce8

Suprisingly the best fix for now is to just remove the duration check.

The issue happened because if the duration is less than 10ms, it is
bucketized into 0 by UKM.
The proper fix would be make the delay longer.

I have considered these options.
(1) change the test into an unit test and use BrowserTaskEnvironment;
this can't be done easily because the whole class heavily depends on
Profile, User, VideoConferenceManager etc.

(2) use a tick_clock and overwrite that in test. this can't be easily
done because we have to pass the clock into each VideoConferenceUkmHelper which is implicitly created and stored in the
client; which means several steps of assignment.

Given the duration is already tested in VideoConferenceUkmHelperTest,
Here in this test we can focus more on the triggering and
bool values of the UKM log.

We can have a more complicated fix later on if required. But not in
this reland.

Original change's description:
> [VC] ukm logging for arc++ apps.
>
> (1) each vc app_id is also mapped to a ukm_helper.
>
> (2) ukm_helper records the start and stop of camera and microphone
>     capturing.
>
> (3) ukm is logged when the app is closed.
>
> Bug: b/266004750
>
> Test="autoninja -C  out/Test chrome/test:browser_tests && out/Test/browser_tests --gtest_filter=VideoConferenceAppServiceClientTest"
>
> Change-Id: I8a7331960de999e48cb7a065a0c64e7804a85684
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4674930
> Commit-Queue: Charles Zhao <charleszhao@chromium.org>
> Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1169660}

Bug: b/266004750
Change-Id: I89f646f85148ad86dc7a45c7a0cca636321f700e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4721131
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Charles Zhao <charleszhao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176382}