CloseWatcher: split up basic.html web platform test
Contrary to its name, basic.html had grown to be a grab-bag of tests.
This takes the first step of splitting it up, creating a separate test
for event properties, and one for what happens when you invoke various
CloseWatcher methods inside event listeners.

event-properties.html is slightly more comprehensive than the test
previously in basic.html, in that it also tests the cancel event's
properties.

inside-event-listeners.html is more exhaustive than the tests previously
in basic.html. It uses the createRecordingCloseWatcher helper to check
the exact events that are fired, and confirm that destroyed close
watchers stay destroyed.

Bug: 1171318
Change-Id: If32f9fbe29482735a82c982c90fe239f6a1edc7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4722573
Commit-Queue: Domenic Denicola <domenic@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176386}