[ios] Merge //base/ios into //base
The target //base/ios was separate because it contained file that
required ARC to be enabled but //base could not enable it because
it also listed some Objective-C++ file that would not compile if
ARC was enabled.

With recent changes that converted all Objective-C++ code in the
codebase to require ARC, this distinction is no longer needed. It
become problematic to have two separate targets because developer
put direct dependencies on //base/ios instead of //base (incorrect)
which causes dependencies issue in some case.

Bug: 733237
Change-Id: I4314a5b80346dd9aefea319ea3718eb8c3990dd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4724428
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176398}