Default Apps: Support stylus capability check on Lacros
The Cursive app preload on ChromeOS is currently restricted to devices
with stylus-enabled touchscreens. This restriction is implemented by
checking ui::DeviceDataManager to determine what touchscreens are
available.

On Lacros, ui::DeviceDataManager is populated by Wayland, and does not
have the full information needed to implement this check. Instead, we
perform the check on Ash side and pass it through DeviceProperties on
crosapi (which already has similar values for other preload properties).

As DeviceDataManager is populated asynchronously, there is a chance that
it might not be ready by the time Lacros starts. I have never seen this
happen in local testing, but we already have UMA in place to detect if
this happens. We could adjust behavior on the Lacros side to not modify
the current preload state if stylus capability is unknown.

In the medium-to-long term, we hope to remove the stylus-only
restriction on the Cursive preload, which would remove the need for this
crosapi parameter.

Bug: 1360294

Change-Id: I85cf6401d03b10d2e97831d0f11f85bcdccef91a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4706205
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176447}