URL: Measure escaped ASCII chars in URL path
Chrome currently unescapes percent-encoded ASCII characters in URL
path. However, this behavior is non-compliant with the URL Standard.

To address this interoperability issue, we need to assess the impact
of fixing it.

I worried about measuring being a hotspot since this can happen when
constructing a GURL or KURL. However, we added a similar UMA for URL
host part [1] in the past. So most likely this will be OK.

[1] https://crrev.com/c/3152695


Bug: 1252531
Change-Id: Ifa1f533fae43c10c81c3bf945844156cc4ca1c34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4713927
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176456}