Browser to ignore global geo switch if PH off
Browser should ignore the cros global geolocation switch if the
PrivacyHub feature is off. If it does not, users may find themselves in
a state where the geolocation is disabled and they have no easy way to
re-enable it.

This patch implements this fully in Ash.
For Lacros we disable the geolocation switch in browser entirely to fix
the user facing problem.
This is not a problem, because geolocation is planned only for a future
version of PrivacyHub, so if it stops working on Lacros, this will not
affect users (unless they use the experimental flag).
A proper fix will be implemented as a change to crosapi later.

BUG=b/293398125
BUG=1464302

Change-Id: I71b3b7756d05cbed7022acb827f797ed64d24345
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4720876
Commit-Queue: Jan Láník <janlanik@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176468}