Rename manual fallback ExplicitlyTriggered metric.
It was decided to change the name of the metric slightly, see [1].

It's generally not recommended to rename metrics. However, this
particular metric was only implemented recently (crrev.com/c/4660421)
and is only relevant for a feature that hasn't started rolling out yet.
(The feature being AutofillShowManualFallbackInContextMenu)

[1] https://docs.google.com/document/d/1Lb346eaee6GfQ6iDbPH44hz_VwyPcpj9Pj-2O2TK9vE/edit?resourcekey=0-hW0oTZZhhiGLiQEu81ptfA

Bug: 1446318
Change-Id: Ia5e9a8a92cba5b861502ad276ae2c6c7824d9b8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4724883
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1176491}