[iOS] Fix crash on account sign out from account settings.
The crash happens in accountsTVC as settingsWillBeDismissed should be
called before dealloc and that is not guaranteed when
ManageSyncSettingsCoordinator gets the notification first.

This CL makes sure it is called before dealloc by having its dismissal
handled by the parent coordinator; ManageSyncSettingsCoordinator and not
by itself.

Bug: 1465416
Change-Id: I229cd5666e4ba8babc784d8a103e4fe777f0f045
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4705347
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Cr-Commit-Position: refs/heads/main@{#1176561}