dpd: enable dpd on non official builds + dcheck_always_on
Similarly to enabling dpd on debug builds, we want to broaden the usage
of DPD by developers. With this CL, we enable DPD on non official builds
that have dcheck_always_on activated, which is a common configuration
for chromium developers. This should also ensure we are not affecting
official builds.

Change-Id: I4921421745dde5dd2967af1c2a217247a731be13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4660451
Commit-Queue: Paul Semel <paulsemel@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176587}